% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipe.R
\name{get_recipe}
\alias{get_recipe}
\title{Get Recipe}
\usage{
get_recipe(cocktail_name)
}
\arguments{
\item{cocktail_name}{A character string specifying the name of the cocktail.}
}
\value{
A data.frame containing the cocktail recipe.
}
\description{
Get the recipe for a specific cocktail by name.
}
