% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_recipes.R
\name{load_recipes}
\alias{load_recipes}
\title{Load Recipes}
\usage{
load_recipes()
}
\value{
A data.frame containing the cleaned cocktail recipes.
}
\description{
Load and clean a database of cocktail recipes.
}
