% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_cocktail.R
\name{suggest_cocktail}
\alias{suggest_cocktail}
\title{Suggest Cocktail}
\usage{
suggest_cocktail(ingredients)
}
\arguments{
\item{ingredients}{A character vector of available ingredients.}
}
\value{
A data.frame containing the suggested cocktail recipes.
}
\description{
Suggest a cocktail recipe based on available ingredients.
}
