% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_random.R
\name{suggest_random}
\alias{suggest_random}
\title{Suggest Random Cocktail}
\usage{
suggest_random(
  category = c("Cocktail", "Shot", "Beer", "Milk / Float / Shake", "Ordinary Drink",
    "Other/Unknown", "Homemade Liqueur", "Punch / Party Drink", "Coffee / Tea",
    "Soft Drink / Soda", "Cocoa")
)
}
\arguments{
\item{category}{A character string specifying the category of the drink (e.g., "Cocktail", "Shot").}
}
\value{
A data.frame containing the randomly selected cocktail recipe.
}
\description{
Suggest a random cocktail based on category.
}
